ResumeRAG: Interactive Resume Retrieval and Generation System/requirements.txt
langchain_community 
tiktoken 
langchain-groq 
langchainhub 
chromadb 
langchain 
langgraph 
langchain_huggingface
langchain
python-dotenv
ipykernel
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
sentence_transformers
langserve
langchain_core
pinecone
pinecone-text
nltk